refactor(cli): move deploy_test.py into tests/cli/ (#899)
